requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.2
requests==2.26.0
urllib3==1.26.6
Werkzeug==2.0.1
